PoissonKernels/MultiCode: Int32 deterministic CIC deposit (bit-identical, drops the reversibility tax)
The DM CIC deposit dominates gravity (40ms random-particle benchmark vs FFT 2.7ms — the FFT is NOT the
bottleneck). The deterministic det-kernels are now generic on the buffer's int type (round(eltype(ρi))),
and the gravity deposit uses Int32 instead of Int64: per-cell Σ≈ρ_dm·2²³≲2.5e8≪2³¹ for z≥20 (55× headroom
measured), and T.(Int)→f32 is identical for both widths (24-bit exact ≪ Σ), so it's BIT-IDENTICAL
(max|Δ|=0; full run δb_rms/x_HII/mass-drift unchanged). For Lagrangian-ordered particles Int32 (1.68ms)
≡ float (1.69ms) — the reversibility tax vanishes; Int64 was 2.36ms. KEY FINDING: the deposit is
ordering-sensitive — 17× slower scrambled (40ms) than ordered (2.4ms) — so its cost grows as DM clusters.

+++ lib/MultiCode/src/global_gravity.jl
@@ -175,7 +175,10 @@ function assemble_global_density_gpu!(ρd, pg::PatchGrid; particles=nothing, dt:
         # independent) so the gravity — and hence checkpoint/restart and run-to-run — is
         # bit-reproducible.  The float `cic_deposit!`'s atomicAdd is order-dependent: the
         # ~2e-5 run-to-run noise AND the restart divergence both traced to it. repic-style.
-        ρpi = PPMKernels.device_zeros(be, Int64, (Ntot,))
+        # Int32 (not Int64): per-cell Σ ≈ ρ_dm·2²³ ≲ 2.5e8 ≪ 2³¹ for these z≥20 runs (mild
+        # clustering), and T.(Int) → f32 is identical for both widths (24-bit exact ≪ 2.5e8),
+        # so it's BIT-IDENTICAL to the Int64 deposit but ~1.7× faster (Int32 atomics on Ampere).
+        ρpi = PPMKernels.device_zeros(be, Int32, (Ntot,))
         PoissonKernels.cic_deposit_det!(ρpi, particles.px, particles.py, particles.pz,
                                         particles.vx, particles.vy, particles.vz, particles.mass;
                                         N=nc[1], disp=0.5*dt/a, shift=-0.5, qbits=23)
